src/requirements.txt
fastapi[all]
uvicorn
pydantic
pydantic-settings
python-dotenv
pytz
pytest
requests
pyswisseph
slowapi